Shorten code, change formatting, edit prints

+++ main.py
@@ -7,40 +7,35 @@ def write_links(comic_number: int):
 
     nf = open("imagelist.txt", "a")
 
+    # Url format for the images has changed multiple times, this part determines which url format to use and writes the link correctly
     if comic_number == 19 or comic_number == 37:
-        # These comic numbers might be unused, no comic are found at these numbers, they are skipped
-        print("Skipping unused comic number #" + str(comic_number) +
-              ". This is normal and not an error. Download continues from next comic...")
+        print("Skipping unused comic number #" + str(comic_number) + ". This is normal and not an error. No comic can be found at this number.")
 
     elif comic_number == 5:
         i = 1
         while i < 6:
-            nf.write("https://assets.nintendo.com/image/upload/w_600,f_auto,q_auto/v1686932568/Microsites/PIKMIN-Portal/comics/00" +
-                     str(comic_number) + "/pikmin-comic-00" + str(comic_number) + "_0" + str(i))
+            nf.write("https://assets.nintendo.com/image/upload/w_600,f_auto,q_auto/v1686932568/Microsites/PIKMIN-Portal/comics/00" + str(comic_number) + "/pikmin-comic-00" + str(comic_number) + "_0" + str(i))
             i += 1
             nf.write("\n")
 
     elif comic_number < 10:
         i = 1
         while i < 6:
-            nf.write("https://assets.nintendo.com/image/upload/w_600,f_auto,q_auto/v1686959530/Microsites/PIKMIN-Portal/comics/00" +
-                     str(comic_number) + "/EN/nint2402-pikmin4-manga00" + str(comic_number) + "_0" + str(i))
+            nf.write("https://assets.nintendo.com/image/upload/w_600,f_auto,q_auto/v1686959530/Microsites/PIKMIN-Portal/comics/00" + str(comic_number) + "/EN/nint2402-pikmin4-manga00" + str(comic_number) + "_0" + str(i))
             i += 1
             nf.write("\n")
 
     elif comic_number >= 10 and comic_number < 30:
         i = 1
         while i < 6:
-            nf.write("https://assets.nintendo.com/image/upload/w_600,f_auto,q_auto/v1688141442/Microsites/PIKMIN-Portal/comics/0" +
-                     str(comic_number) + "/EN/nint2402-pikmin4-manga0" + str(comic_number) + "_0" + str(i))
+            nf.write("https://assets.nintendo.com/image/upload/w_600,f_auto,q_auto/v1688141442/Microsites/PIKMIN-Portal/comics/0" + str(comic_number) + "/EN/nint2402-pikmin4-manga0" + str(comic_number) + "_0" + str(i))
             i += 1
             nf.write("\n")
 
     elif comic_number >= 30:
         i = 1
         while i < 6:
-            nf.write("https://assets.nintendo.com/image/upload/w_600,f_auto,q_auto/v1694014678/Microsites/PIKMIN-Portal/comics/" +
-                     str(comic_number) + "/nint2402-pikmin4-manga0" + str(comic_number) + "_0" + str(i))
+            nf.write("https://assets.nintendo.com/image/upload/w_600,f_auto,q_auto/v1694014678/Microsites/PIKMIN-Portal/comics/" + str(comic_number) + "/nint2402-pikmin4-manga0" + str(comic_number) + "_0" + str(i))
             i += 1
             nf.write("\n")
 
@@ -69,34 +64,25 @@ def generate_image_list():
         case _:
             comic_amount = 50
 
-    # Create text file
-    nf = open("imagelist.txt", "w")
+    # Create text file if it does not exist
+    if not os.path.exists("imagelist.txt"):
+        nf = open("imagelist.txt", "w")
+        nf.close()
 
-    # Generate images
+    # Loop through panel numbers and send them one by one to write_links function
     i = 1
     while i <= int(comic_amount):
-        if i <= 5:
-            write_links(i)
-        if i == 6:
-            write_links(i)
-        if i > 6 and i < 10:
-            write_links(i)
-        if i >= 10 and i < 30:
-            write_links(i)
-        if i >= 30:
-            write_links(i)
+        write_links(i)
         i += 1
-
-    nf.close()
 
     input("Task finished. Press Enter to start download...")
 
 
 def download_images():
 
+    target_folder = 'comic/'
     current_page = 1
     image_name = str("Pik4-comic-1")
-    target_folder = 'comic/'
 
     # Create comic folder if it does not exist
     if not os.path.exists(target_folder):
@@ -118,8 +104,7 @@ def download_images():
             # Save image to the folder
             with open(f'{target_folder}{file_name}', 'wb') as file:
                 file.write(image_link.content)
-                print(style.GREEN +
-                      f'Image download #' + str(current_page) + ' finished with file name ' + file_name + '.' + style.RESET)
+                print(style.GREEN + f'Image download #' + str(current_page) + ' finished with file name ' + file_name + '.' + style.RESET)
 
             current_page += 1
             image_name = str("Pik4-comic-")
@@ -129,7 +114,7 @@ def download_images():
 
         # Close text file
         f.close()
-        print(style.RESET + 'Download complete, check your "comic" folder.')
+        print(style.RESET + 'Download complete, images are in "comic" folder.')
 
     else:
         print(
@@ -146,13 +131,11 @@ if __name__ == '__main__':
         RED = '\033[31m'
         GREEN = '\033[32m'
         YELLOW = '\033[33m'
-        UNDERLINE = '\033[4m'
         RESET = '\033[0m'
 
     # Choice of download method
-    print(style.YELLOW +
-          "\nChoose method by typing the number and pressing Enter:" + style.RESET)
-    print("1. Automatic image list generation + download images")
+    print(style.YELLOW + "\nChoose method by typing the number and pressing Enter:" + style.RESET)
+    print("1. (Recommended) Automatic image list generation + download images")
     print("2. Use existing image list + download images")
     list_choice = input()
 
@@ -169,7 +152,6 @@ if __name__ == '__main__':
 
         case _:
             print(style.RED + "Error #1: Invalid input." + style.RESET)
-            print(style.RESET)
             exit()
 
     print(style.RESET + "Closing program...")
